requirements.txt
streamlit
pandas
feedparser
scikit-learn
google-generativeai
requests